refactor(grpc): extract shared compute_backoff helper to avoid duplication
`compute_backoff` and its six unit tests were identical in both
`client.rs` and `rpc_retry.rs`. Move them into a new private
`backoff` module (`backoff.rs`) and update both callers to use
`crate::backoff::compute_backoff`. No behaviour change.

Resolves review comments from @Bechma on PR #825.

Made-with: Cursor
Made-with: Cursor
Made-with: Cursor
